+++ turses/api/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+"""
+turses.api
+~~~~~~~~~~
+
+This module contains the Twitter API implementations.
+"""