+++ turses/cli.py
@@ -1,12 +1,16 @@
-###############################################################################
-#                               coding=utf-8                                  #
-#            Copyright (c) 2012 turses contributors. See AUTHORS.             #
-#         Licensed under the GPL License. See LICENSE for full details.       #
-###############################################################################
+# -*- coding: utf-8 -*-
+
+"""
+turses.cli
+~~~~~~~~~~
+
+This module contains the logic to launch `turses` with a curses interface.
+"""
 
 from .util import parse_arguments
 from .config import Configuration
-from .controller import Turses
+from .controller import CursesController
+from .constant import palette
 from .ui.curses import CursesInterface
 
 
@@ -14,6 +18,8 @@ def main():
     try:
         configuration = Configuration(parse_arguments())
         ui = CursesInterface()
-        Turses(configuration, ui)
+        CursesController(palette, 
+                         configuration, 
+                         ui)
     except KeyboardInterrupt:
         exit(0)