+++ turses/ui/__init__.py
@@ -1,5 +1,8 @@
-###############################################################################
-#                               coding=utf-8                                  #
-#            Copyright (c) 2012 turses contributors. See AUTHORS.             #
-#         Licensed under the GPL License. See LICENSE for full details.       #
-###############################################################################
+# -*- coding: utf-8 -*-
+
+"""
+turses.ui
+~~~~~~~~~
+
+This module contains the UI frontends to `turses`.
+"""