+++ HISTORY.rst
@@ -1,3 +1,9 @@
+0.0.10
+-----
+- unread count
+- mark all tweets in active timeline as read
+- fix (again) a bug with mouse events
+
 0.0.9
 -----
 - compose tweet with same hashtags as the focused status